fix: import `overridable-registry.js` in `records_list` macro
* The `records_list` macro was importing `invenio-app-rdm-frontpage.js`
early, before any other component-specific JS was imported. This is
necessary, as the macro can be used in a range of places, so we need to
make sure its associated JS is imported wherever it is used.

* Since we were importing this JS before `overridable-registry`, the
loop to add the overriden components to the singleton store
(https://github.com/inveniosoftware/invenio-app-rdm/blob/979d6ab4c84fffaa8c70b2b32a40efac179bfd98/invenio_app_rdm/theme/assets/semantic-ui/js/invenio_app_rdm/overridableRegistry/index.js#L11-L13)
was being run _after_ the frontpage JS ran the `getAll` function on the
store. This meant user-specified overrides were not included in the
object (which is a deep-cloned value rather than a reference), so failed
to get rendered correctly on the frontpage.

* This PR adds the `overridable-registry.js` import to the
`records_list` macro. Even though this will result in a page including
the import multiple times, the import behaves as a singleton so it won't
actually be imported more than once. This way, we ensure the import is
in the correct location regardless of where the macro is used.